service-api/requirements.txt
flask
tzlocal
requests
gunicorn
flask-cors
psycopg2-binary